SCHED-304: Fixing files.
